+++ core-services/beneficiary-idgen/src/main/java/org/egov/id/service/EnrichmentService.java
@@ -19,7 +19,7 @@ import java.util.Map;
 public class EnrichmentService {
 
 
-    public void update(List<IdRecord> validIdRecords, IdRecordBulkRequest request) throws Exception {
+    public void update(List<IdRecord> validIdRecords, IdRecordBulkRequest request) {
         log.info("starting the enrichment for update idRecords");
         Map<String, IdRecord> iMap = getIdToObjMap(validIdRecords);
         log.info("enriching individuals for update");
@@ -33,11 +33,7 @@ public class EnrichmentService {
                     idRecord.setStatus(status);
                 }
         );
-        log.info("starting the enrichment for update idRecords");
-        Map<String, IdRecord> iMap = getIdToObjMap(validIdRecords);
-        log.info("enriching individuals for update");
-        enrichForUpdate(iMap, request);
-        log.info("completed the enrichment for update idRecords");
+        update(validIdRecords, request);
     }
 
 }