+++ health-services/libraries/health-services-models/src/main/java/org/egov/common/models/idgen/IDPoolCreationResult.java
@@ -1,28 +1,21 @@
 package org.egov.common.models.idgen;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import lombok.ToString;
 
+@Builder
 @Getter
 @Setter
-@NoArgsConstructor
 @AllArgsConstructor
-@ToString
-public class BatchRequest {
-
-
-    @JsonProperty("batchSize")
-    @NotNull
-    @Positive
-    private Integer batchSize;
-
+@NoArgsConstructor
+public class IDPoolCreationResult {
     @JsonProperty("tenantId")
-    @NotNull
     private String tenantId;
+    @JsonProperty("message")
+    private String message;
 }
+