+++ health-services/libraries/health-services-models/src/main/java/org/egov/common/models/idgen/IDPoolGenerationResponse.java
@@ -34,6 +34,6 @@ public class IDPoolGenerationResponse {
     @JsonProperty("IdCreationResponse")
     @NotNull
     @Valid
-    private List<Map<String,String>> idCreationResponse;
+    private List<IDPoolCreationResult> idPoolCreateResponses;
 
 }